+++ lib/crunchbase/models/entity.rb
@@ -24,8 +24,6 @@ module Crunchbase
       def setup_relationships(object, request_card_ids, response_cards)
         request_card_ids.each do |card_id|
           card_data = response_cards.dig(card_id)
-          next if card_data.nil? || card_data.empty?
-
           card_model = card_model_mappings[card_id]
           card_objects = if card_data.is_a?(Array)
                            card_data.each_with_object([]) do |data, objects|
@@ -33,7 +31,7 @@ module Crunchbase
                              objects << dynamic_attributes(new_card_instance, new_card_instance.basis_fields, data)
                            end
                          else
-                           dynamic_attributes(card_model.new, extract_fields, data)
+                           card_data.nil? ? nil : dynamic_attributes(card_model.new, extract_fields, data)
                          end
 
           dynamic_define_method(object, card_id, card_objects)
@@ -48,21 +46,21 @@ module Crunchbase
         {
           'investments' => Crunchbase::Models::Investment,
           'raised_investments' => Crunchbase::Models::Investment,
-          # 'participated_funds' => Crunchbase::Models::Fund,
-          'child_organizations' => Crunchbase::Models::Organization,
           'participated_investments' => Crunchbase::Models::Investment,
-          # 'investors' => Crunchbase::Models::Investor,
+          'participated_funds' => Crunchbase::Models::Fund,
+          'raised_funds' => Crunchbase::Models::Fund,
+          'child_organizations' => Crunchbase::Models::Organization,
           'parent_organization' => Crunchbase::Models::Organization,
+          # 'investors' => Crunchbase::Models::Investor,
           'raised_funding_rounds' => Crunchbase::Models::FundingRound,
+          'participated_funding_rounds' => Crunchbase::Models::FundingRound,
           'ipos' => Crunchbase::Models::Ipo,
           # 'event_appearances' => Crunchbase::Models::EventAppearance,
-          # 'raised_funds' => Crunchbase::Models::Fund,
           'acquiree_acquisitions' => Crunchbase::Models::Acquisition,
           # 'parent_ownership' => Crunchbase::Models::Ownership,
+          # 'child_ownerships' => Crunchbase::Models::Ownership,
           # 'jobs' => Crunchbase::Models::Job,
           # 'founders' => Crunchbase::Models::Founder,
-          # 'child_ownerships' => Crunchbase::Models::Ownership,
-          'participated_funding_rounds' => Crunchbase::Models::FundingRound,
           'press_references' => Crunchbase::Models::PressReference
           # 'headquarters_address' => Crunchbase::Models::Address,
         }